+++ gulp/tasks/clean.js
@@ -1,0 +1,5 @@
+import del from 'del';
+
+export const clean = () => {
+  return del('build');
+}
