+++ gulp/tasks/copy.js
@@ -1,0 +1,11 @@
+export const copy = (done) => {
+  gulp.src([
+    'source/fonts/*.{woff,woff2}',
+    'source/*.ico',
+    'source/manifest.webmanifest'
+  ], {
+    base: 'source'
+  })
+    .pipe(gulp.dest('build'));
+  done();
+}
