+++ gulp/tasks/copyImages.js
@@ -1,0 +1,4 @@
+export const copyImages = () => {
+  return gulp.src('source/img/**.*.{jpg,png}')
+    .pipe(gulp.dest('build/img'));
+}
