+++ gulp/tasks/html.js
@@ -1,0 +1,7 @@
+import htmlmin from 'gulp-htmlmin';
+
+export const html = () => {
+  return gulp.src('source/*.html')
+    .pipe(htmlmin({ collapseWhitespace: true }))
+    .pipe(gulp.dest('build'));
+}
