+++ gulp/tasks/optimizeImages.js
@@ -1,0 +1,7 @@
+import squoosh from 'gulp-libsquoosh';
+
+export const optimizeImages = () => {
+  return gulp.src('source/img/**/*.{jpg,png}')
+    .pipe(squoosh())
+    .pipe(gulp.dest('build/img'));
+}
