+++ gulp/tasks/optimizeSvg.js
@@ -1,0 +1,7 @@
+import svgo from 'gulp-svgmin';
+
+export const optimizeSvg = () => {
+  return gulp.src(['source/img/*.svg', '!source/img/icons/*.svg'])
+    .pipe(svgo())
+    .pipe(gulp.dest('build/img'));
+}
