+++ gulp/tasks/scripts.js
@@ -1,0 +1,7 @@
+import terser from 'gulp-terser';
+
+export const scripts = () => {
+  return gulp.src('source/js/*.js')
+    .pipe(terser())
+    .pipe(gulp.dest('build/js'));
+}
