+++ gulp/tasks/server.js
@@ -1,0 +1,11 @@
+export const server = (done) => {
+  browser.init({
+    server: {
+      baseDir: 'build'
+    },
+    cors: true,
+    notify: false,
+    ui: false,
+  });
+  done();
+}
