+++ gulpfile.js
@@ -1,122 +1,24 @@
 import gulp from 'gulp';
-import plumber from 'gulp-plumber';
-import sass from 'gulp-dart-sass';
-import postcss from 'gulp-postcss';
-import csso from 'postcss-csso';
-import autoprefixer from 'autoprefixer';
 import browser from 'browser-sync';
-import htmlmin from 'gulp-htmlmin';
-import terser from 'gulp-terser';
-import squoosh from 'gulp-libsquoosh';
-import svgo from 'gulp-svgmin';
-import svgstore from 'gulp-svgstore';
-import rename from 'gulp-rename';
-import del from 'del';
+import { clean } from './gulp/tasks/clean.js';
+import { copy } from './gulp/tasks/copy.js';
+import { copyImages } from './gulp/tasks/copyImages.js';
+import { optimizeImages } from './gulp/tasks/optimizeImages.js';
+import { html } from './gulp/tasks/html.js';
+import { styles } from './gulp/tasks/styles.js';
+import { scripts } from './gulp/tasks/scripts.js';
+import { createWebp } from './gulp/tasks/createWebp.js';
+import { optimizeSvg } from './gulp/tasks/optimizeSvg.js';
+import { createSvgSprite } from './gulp/tasks/createSvgSprite.js';
+import { server } from './gulp/tasks/server.js';
 
-// Html
+global.gulp = gulp;
+global.browser = browser;
 
-const html = () => {
-  return gulp.src('source/*.html')
-    .pipe(htmlmin({ collapseWhitespace: true }))
-    .pipe(gulp.dest('build'));
-}
-
-// Styles
-
-export const styles = () => {
-  return gulp.src('source/sass/style.scss', { sourcemaps: true })
-    .pipe(plumber())
-    .pipe(sass().on('error', sass.logError))
-    .pipe(postcss([
-      autoprefixer(),
-      csso()
-    ]))
-    .pipe(rename('style.min.css'))
-    .pipe(gulp.dest('build/css', { sourcemaps: '.' }))
-    .pipe(browser.stream());
-}
-
-// JavaScript
-
-const scripts = () => {
-  return gulp.src('source/js/*.js')
-    .pipe(terser())
-    .pipe(gulp.dest('build/js'));
-}
-
-// Images
-
-const optimizeImages = () => {
-  return gulp.src('source/img/**/*.{jpg,png}')
-    .pipe(squoosh())
-    .pipe(gulp.dest('build/img'));
-}
-
-const copyImages = () => {
-  return gulp.src('source/img/**.*.{jpg,png}')
-    .pipe(gulp.dest('build/img'));
-}
-
-// WebP
-
-const createWebp = () => {
-  return gulp.src('source/img/**.*.{jpg,png}')
-    .pipe(squoosh({
-      webp: {}
-    }))
-    .pipe(gulp.dest('build/img'));
-}
-
-// SVG
-
-const optimizeSvg = () => {
-  return gulp.src(['source/img/*.svg', '!source/img/icons/*.svg'])
-    .pipe(svgo())
-    .pipe(gulp.dest('build/img'));
-}
-
-const createSvgSprite = () => {
-  return gulp.src('source/img/icons/*.svg')
-    .pipe(svgo())
-    .pipe(svgstore({
-      inlineSvg: true
-    }))
-    .pipe(rename('sprite.svg'))
-    .pipe(gulp.dest('build/img'));
-}
-
-// Copy
-
-const copy = (done) => {
-  gulp.src([
-    'source/fonts/*.{woff,woff2}',
-    'source/*.ico',
-    'source/manifest.webmanifest'
-  ], {
-    base: 'source'
-  })
-    .pipe(gulp.dest('build'));
-  done();
-}
-
-// Clean
-
-const clean = () => {
-  return del('build');
-}
-
-// Server
-
-const server = (done) => {
-  browser.init({
-    server: {
-      baseDir: 'build'
-    },
-    cors: true,
-    notify: false,
-    ui: false,
-  });
-  done();
+export const watcher = () => {
+  gulp.watch('source/sass/**/*.scss', gulp.series(styles));
+  gulp.watch('source/js/*.js', gulp.series(scripts));
+  gulp.watch('source/*.html', gulp.series(html, browser.reload));
 }
 
 export const build = gulp.series(
@@ -131,14 +33,6 @@ export const build = gulp.series(
     optimizeSvg,
     createSvgSprite
 ));
-
-// Watcher
-
-const watcher = () => {
-  gulp.watch('source/sass/**/*.scss', gulp.series(styles));
-  gulp.watch('source/js/*.js', gulp.series(scripts));
-  gulp.watch('source/*.html', gulp.series(html, browser.reload));
-}
 
 export default gulp.series(
   clean,
