+++ package.json
@@ -14,6 +14,7 @@
     "gulp-plumber": "1.2.1",
     "gulp-postcss": "9.0.1",
     "gulp-rename": "^2.0.0",
+    "gulp-stacksvg": "^1.1.0",
     "gulp-svgmin": "^4.1.0",
     "gulp-svgstore": "8.0.0",
     "gulp-terser": "^2.1.0",
